Refactor googlePlayGameServices
