--- android/src/car/superfun/game/multiplayer/MultiplayerManager.java
@@ -1,5 +1,0 @@
-package car.superfun.game.multiplayer;
-
-public class MultiplayerManager {
-
-}
